画面遷移と再レンダリング及び画面レイアウト

+++ diary-app-frontend/src/components/DiaryDetail.js
@@ -3,12 +3,13 @@ import { useParams,useNavigate} from 'react-router-dom';
 import diaryService from "../services/diaryService";
 
 //詳細画面を表示するコンポーネントを作成
-const DiaryDetail = () => {
+const DiaryDetail = ({ refreshDiaries }) => {
   //特定の日記データをセット
   const [diary, setDiary] = useState(null);
   //ルートパラメーターからidを取得し、取得したオブジェクトを数値にする
   const params = useParams();
   const diaryId = parseInt(params["id"]);
+  
   useEffect(() => {
     if (diaryId) {
       diaryService.getDiaryDetail(diaryId).then(response => {
@@ -20,35 +21,48 @@ const DiaryDetail = () => {
     }
   }, [diaryId]);
 
-    //削除ボタンを押下したときの処理
-    const deleteDiary = (id) => {
-      diaryService.deleteDiary(id).then(() => {
-        alert("指定された日記を削除しました");
-        //TODO:一覧画面に遷移する処理をここに
-      })
-        .catch(err => {
-          console.error("Error delete diary: ", err);
-          alert("指定された日記の削除に失敗しました");
-      })
+    //日記の削除処理
+    const deleteDiary = (id, diaries) => {
+      const result =  window.confirm("削除します。本当に宜しいですか？");
+      if (result) {
+        diaryService.deleteDiary(id).then(() => {
+          alert("指定された日記を削除しました");
+          refreshDiaries();
+          navigate("/diaries"); // 削除完了後に一覧画面に遷移
+        })
+          .catch(err => {
+            console.error("Error delete diary: ", err);
+            alert("指定された日記の削除に失敗しました");
+        })
+      }
     }
-  
+
   const navigate = useNavigate();
-  //日記が見つかった場合と見つからなかった場合で返す処理を分ける
+
+  //指定された日記が存在していれば、詳細画面を表示し、そうでなければページにメッセージを表示
   if (diary) {
   return (
-      <div>
+      <div className="diaryBox diaryDetail">
         <h1>日記の詳細</h1>
-        <ul>
-          <li>{diary.title}</li>
-          <li>{diary.date}</li>
-          <li>{diary.content}</li>
-          <button onClick = {()=>navigate(`/diaries/edit/${diaryId}`)}>編集</button>
-        <button onClick={() => deleteDiary(diaryId)}>削除</button>
-        </ul>
+      <ul className="contentsBox">
+        <div className="contents">
+          <div className="contentsHead">
+            <li>{diary.title}</li>
+            <li>{diary.date}</li>
+          </div>
+          <li >{diary.content}</li>
+        </div >
+        <div className="buttonArea">
+          <button className="successBtn"onClick = {()=>navigate(`/diaries/edit/${diaryId}`)}>編集</button>
+          <button className="deleteBtn" onClick={() => deleteDiary(diaryId)}>削除</button>
+        </div>
+      </ul>
       </div>
     );
   } else {
-    <div>Loading...</div>;
+    <div>
+      <p>この日記は存在しない、または削除されたため、表示できません。</p>
+    </div>;
   }
 }
 export default DiaryDetail;
